utils file for lineAreaGraph
